Criação do método Exist para validação se cliente existe durante o update e delete. ŔEFATORAR

+++ SuspirarDoces.Application/Services/ClientService.cs
@@ -42,14 +42,36 @@ namespace SuspirarDoces.Application.Services
 
         public void Remove(int? id)
         {
-            var client = _clientRepository.GetById(id).Result;
-            _clientRepository.Remove(client);
+            if (Exist(id))
+            {
+                var client = _clientRepository.GetById(id).Result;
+                _clientRepository.Remove(client);
+            }
+            else
+            {
+                throw new Exception("Cliente informado não existe");
+            }
         }
 
         public void Update(ClientViewModel entity)
         {
-            var client = _mapper.Map<Cliente>(entity);
-            _clientRepository.Update(client);
+            var exist = Exist(entity.Id);
+            if(exist == false)
+            {
+                throw new Exception("Cliente informado não existe");
+            }
+            else
+            {
+                var client = _mapper.Map<Cliente>(entity);
+                _clientRepository.Update(client);
+            }
+        }
+
+        public bool Exist(int? id)
+        {
+            var client = _clientRepository.GetById(id).Result;
+            if (client == null) return false;
+            return true;
         }
     }
 }
